+++ api/lib/application/authentication/oidc/oidc-controller.js
@@ -1,30 +1,6 @@
 import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
 import { oidcAuthenticationServiceRegistry, usecases } from '../../../domain/usecases/index.js';
 
-const reconcileUser = async function (
-  request,
-  h,
-  dependencies = {
-    oidcAuthenticationServiceRegistry,
-  },
-) {
-  const { identityProvider, authenticationKey } = request.deserializedPayload;
-
-  await dependencies.oidcAuthenticationServiceRegistry.loadOidcProviderServices();
-  await dependencies.oidcAuthenticationServiceRegistry.configureReadyOidcProviderServiceByCode(identityProvider);
-
-  const oidcAuthenticationService = dependencies.oidcAuthenticationServiceRegistry.getOidcProviderServiceByCode({
-    identityProviderCode: identityProvider,
-  });
-
-  const result = await usecases.reconcileOidcUser({
-    authenticationKey,
-    oidcAuthenticationService,
-  });
-
-  return h.response({ access_token: result.accessToken, logout_url_uuid: result.logoutUrlUUID }).code(200);
-};
-
 const reconcileUserForAdmin = async function (
   request,
   h,
@@ -53,7 +29,6 @@ const reconcileUserForAdmin = async function (
 };
 
 const oidcController = {
-  reconcileUser,
   reconcileUserForAdmin,
 };
