+++ api/src/identity-access-management/application/oidc-provider/oidc-provider.controller.js
@@ -1,4 +1,5 @@
 import { BadRequestError, UnauthorizedError } from '../../../../lib/application/http-errors.js';
+import { oidcAuthenticationServiceRegistry } from '../../../../lib/domain/usecases/index.js';
 import * as oidcSerializer from '../../../../lib/infrastructure/serializers/jsonapi/oidc-serializer.js';
 import { usecases } from '../../domain/usecases/index.js';
 import * as oidcProviderSerializer from '../../infrastructure/serializers/jsonapi/oidc-identity-providers.serializer.js';
@@ -134,12 +135,44 @@ async function getRedirectLogoutUrl(request, h) {
 }
 
 /**
+ * @callback reconcileUser
+ * @param request
+ * @param h
+ * @param dependencies
+ * @return {Promise<{access_token: string, logout_url_uuid: string}>}
+ */
+async function reconcileUser(
+  request,
+  h,
+  dependencies = {
+    oidcAuthenticationServiceRegistry,
+  },
+) {
+  const { identityProvider, authenticationKey } = request.deserializedPayload;
+
+  await dependencies.oidcAuthenticationServiceRegistry.loadOidcProviderServices();
+  await dependencies.oidcAuthenticationServiceRegistry.configureReadyOidcProviderServiceByCode(identityProvider);
+
+  const oidcAuthenticationService = dependencies.oidcAuthenticationServiceRegistry.getOidcProviderServiceByCode({
+    identityProviderCode: identityProvider,
+  });
+
+  const result = await usecases.reconcileOidcUser({
+    authenticationKey,
+    oidcAuthenticationService,
+  });
+
+  return h.response({ access_token: result.accessToken, logout_url_uuid: result.logoutUrlUUID }).code(200);
+}
+
+/**
  * @typedef {Object} OidcProviderController
  * @property {authenticateOidcUser} authenticateOidcUser
  * @property {findUserForReconciliation} findUserForReconciliation
  * @property {getAuthorizationUrl} getAuthorizationUrl
  * @property {getIdentityProviders} getIdentityProviders
  * @property {getRedirectLogoutUrl} getRedirectLogoutUrl
+ * @property {reconcileUser} reconcileUser
  */
 export const oidcProviderController = {
   authenticateOidcUser,
@@ -148,4 +181,5 @@ export const oidcProviderController = {
   getAuthorizationUrl,
   getIdentityProviders,
   getRedirectLogoutUrl,
+  reconcileUser,
 };
