+++ api/tests/unit/application/authentication/oidc/oidc-controller_test.js
@@ -14,36 +14,6 @@ describe('Unit | Application | Controller | Authentication | OIDC', function () 
     };
   });
 
-  describe('#reconcileUser', function () {
-    it('should call use case and return the result', async function () {
-      // given
-      const request = {
-        deserializedPayload: {
-          identityProvider: 'OIDC',
-          authenticationKey: '123abc',
-        },
-      };
-
-      const dependencies = {
-        oidcAuthenticationServiceRegistry: oidcAuthenticationServiceRegistryStub,
-      };
-      sinon.stub(usecases, 'reconcileOidcUser').resolves({
-        accessToken: 'accessToken',
-        logoutUrlUUID: 'logoutUrlUUID',
-      });
-
-      // when
-      const result = await oidcController.reconcileUser(request, hFake, dependencies);
-
-      // then
-      expect(oidcAuthenticationServiceRegistryStub.loadOidcProviderServices).to.have.been.calledOnce;
-      expect(
-        oidcAuthenticationServiceRegistryStub.configureReadyOidcProviderServiceByCode,
-      ).to.have.been.calledWithExactly(identityProvider);
-      expect(result.source).to.deep.equal({ access_token: 'accessToken', logout_url_uuid: 'logoutUrlUUID' });
-    });
-  });
-
   describe('#reconcileUserForAdmin', function () {
     it('calls use case and return the result', async function () {
       // given
